% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ising_ncon.R
\name{get_ising_ncon}
\alias{get_ising_ncon}
\title{Ising model unbiased normalizing constant}
\usage{
get_ising_ncon(beta, k = 0, m = 1, size, max_iterations = Inf)
}
\arguments{
\item{beta}{inverse temperatures}

\item{k}{An integer at which to start computing the unbiased estimator}

\item{m}{A time horizon: the chains are sampled until the maximum between m and the meeting time}

\item{size}{grid size}
}
\value{
A list with
\itemize{

\item mcmcestimator: an MCMC estimator computed on the first chain, from step k to m

\item correction: the bias correction term

\item uestimator: unbiased estimator, equal to the sum of mcmcestimator and correction

\item meetingtime: the meeting time; equal to Inf if while loop was interrupted

\item iteration: final iteration; could be equal to m, to meetingtime, or to max_iterations

\item finished: state of meeting
}
}
\description{
Sample an "time-averaged" unbiased estimator of the expected nature statistics in Ising model
}
