% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mh_kernels_cut.R
\name{get_kernels}
\alias{get_kernels}
\title{Get kernels for cut distribution}
\usage{
get_kernels(theta1, Sigma_proposal, init_mean, init_Sigma)
}
\arguments{
\item{theta1}{posterior mean under first model}

\item{Sigma_proposal}{covariance of the Normal random walk proposal}

\item{init_mean}{initial mean of the second model}

\item{init_Sigma}{initial covariance of the second model}
}
\value{
A list containing the keys
\code{target}, \code{single_kernel}, \code{coupled_kernel}, \code{rinit}
}
\description{
This function takes a sampled parameter(in the first model), initial status
and a covariance matrix for a Normal random walk proposal, and returns a list containing
the log-pdf target distribution \code{target} as well as \code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled MH kernels.
}
