% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_unbiasedmlmc.R
\name{sample_unbiasedmlmc}
\alias{sample_unbiasedmlmc}
\title{Unbiased ML_MCMC estimators}
\usage{
sample_unbiasedmlmc(g, samples, p)
}
\arguments{
\item{g}{A deterministic function}

\item{samples}{i.i.d samples with expectation equal to E(X)}

\item{p}{parameter of Geometric distribution}
}
\value{
one unbiased estimator whose expectation is g(E(X))
}
\description{
Takes in a density function and a deterministic
function and gives k unbiased estimators with expectation equal to g(E(X))
}
