% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_ising_nstat.R
\name{sampling_ising_nstat}
\alias{sampling_ising_nstat}
\title{sampling unbiased estimators of nature statistics of Ising model}
\usage{
sampling_ising_nstat(betas, k = 0, m = 1, nrep, size)
}
\arguments{
\item{k}{An integer at which to start computing the unbiased estimator}

\item{m}{A time horizon: the chains are sampled until the maximum between m and the meeting time}

\item{nrep}{number of samples needed}

\item{size}{grid size}

\item{beta}{inverse temperatures}
}
\value{
a n*d matrix, where d is the dimension of X and n is the number of samples
}
\description{
Takes in information of Ising model and gives several
unbiased estimators for the expected normalizing constant
}
